# Speed improvements in loading time to Rubi Integration module