… sympy/integrals/rubi/rubi.py