#13241

```python
In [1]: %time from sympy.integrals.rubi.rubi import rubi_integrate
CPU times: user 2.67 s, sys: 96.7 ms, total: 2.77 s
Wall time: 2.81 s
```

* Rubi now uses `.pickle` file to directly load the rules instead of constructing the discrimination net of all the rules(which was time consuming). The pickled file is placed at `sympy/integrals/rubi/`. 

* I have pickled the rules using `dill` module.