@gxyd 2017-09-24
Has this resulted in reduction of importing time? You can paste the timings of import, just like Ondřej did.

BTW I am getting different timings from what @certik posted:
```
In [1]: %time from sympy.integrals.rubi.rubi import rubi_integrate
CPU times: user 12 ms, sys: 0 ns, total: 12 ms
Wall time: 9.92 ms
```