@parsoyaarihant 2017-09-24
It is much faster on my computer now than before:
### Without .pickle file
```python
In [1]: %time from sympy.integrals.rubi.rubi import rubi_integrate
CPU times: user 3min 22s, sys: 1.32 s, total: 3min 23s
Wall time: 3min 26s
```

### With .pickle file
```python
In [1]: %time from sympy.integrals.rubi.rubi import rubi_integrate
CPU times: user 2.67 s, sys: 96.7 ms, total: 2.77 s
Wall time: 2.81 s
```