@parsoyaarihant 2017-09-24
@Upabjojr , I have picked the rules using `dill` module. `dill` makes it possible to pickle `lambda` functions. 

We will need `dill` to load the rules too.