@Upabjojr 2017-09-24
> We will need dill to load the rules too.

OK, though I still believe that the best solution would be to make the lambdas full functions, so that we can avoid using _dill_.