@Upabjojr 2017-09-24
How did you use `dill`, I don't see it in the code.

As another issue: I'm strongly opposed to include binary files into SymPy. If needed, I would rather suggest to include the code to generate the binary file, but not the binary file itself.