@Upabjojr 2017-09-24
Furthermore, we still need to do some work with the rules, in many cases they give the wrong results. I believe we could have another GSoC next year to finish RUBI.