@Upabjojr 2017-09-24
Including binary files doesn't allow to compare the versions and is likely to burst the size of the development branch.

I remember experiencing something similar in the past, I've seen a development branch requiring around 100 GB of space, 99.99% used by binary files regenerated for every commit.