@parsoyaarihant 2017-09-24
> How did you use dill, I don't see it in the code.


I have used `dill` to pickle the matcher. That will need to be installed to load the rules. I thought `pickle` module would be enough but it isn't. 

The code to generate the binary file is really trivial:
```python
import dill
from sympy.integrals.rubi.rubi import rubi_object
rubi = rubi_object()
file_Name = 'pickled_rubi_rules.p'
fileObject = open(file_Name, 'wb')
dill.dump(rubi, fileObject)
fileObject.close()
```
Shall I add this in documentation instead?