@parsoyaarihant 2017-09-24
I was going to start working on testing the rules again after getting this work. I wanted to improve loading time to make testing easier.