@Upabjojr 2017-09-24
You can include the generator of the pickled file. There's no way we are going to include a binary pickle file.