@Upabjojr 2017-09-24
> Shall I add this in documentation?

You can add that code directly to the SymPy code, just use `import_module` for _dill_.