@Upabjojr 2017-09-24
By the way, in order to match `exp(x)` I believe we need to transform `exp(x)` into `Pow(S.Exp1, x, evaluate=False)`, as a preprocessor to the integral.