@asmeurer 2017-09-24
How big is the binary file? GitHub doesn't show the filesize.