@asmeurer 2017-09-24
You can try cloudpickle. I believe unlike dill, it generates a valid pickle format, so you only need it to serialize, not to deserialize.