@ylemkimon 2017-09-25
@asmeurer It saids 6.04 MB at https://github.com/parsoyaarihant/sympy/blob/835355c3135b4c3ce2933b8c868c27502c8f2fb6/sympy/integrals/rubi/pickled_rubi_rules.p .