@Upabjojr 2017-09-25
I will **not** merge any PR containing binary files. The big problem is that every time you modify the rules, the binary file is likely to be noticeably different, and _git_ will probably have to store a whole new blog.