@certik 2017-09-25
Yes, please don't merge any binary files into the sympy git repository. For now, why not to create a script that generates this binary blob, and users can (optionally) run it themselves on their machines?