@Upabjojr 2017-09-25
@parsoyaarihant a nice idea would be to generate the pickled file as a background process, I'd use `multiprocessing` for example:

```python
import multiprocessing

def function_generating_blob(...):
    ...

p = multiprocessing.Process(target=function_generating_blob, args=(arg1, arg2, ...))
p.start()
```
(hope I wrote everything correctly) This should start the function under _target_ as a separate process in the background and allow you to continue to use the current Python session. This could be a loading strategy.