@bjodah 2017-09-30
Another +1 for not checking in any generated binaries in the repo. If it's useful in the source distribution it could be made to be generated by the `sdist` command in `setup.py`. On that note, I've always wondered why github doesn't provide a user defined warning limit on total size of PR...