@Upabjojr 2017-09-30
>  If it's useful in the source distribution it could be made to be generated by the sdist command in setup.py.

I think our aim is to generate the decision tree, pickled files can be useful for development purposes.