@Upabjojr 2019-09-29
No binary files into SymPy please. This PR should definitely be closed.